+++ govtool/frontend/src/components/organisms/CreateGovernanceActionSteps/StorageInformation.tsx
@@ -10,8 +10,9 @@ import {
   useScreenDimension,
 } from "@hooks";
 import { Step } from "@molecules";
-import { BgCard, ControlledField } from "@organisms";
+import { BgCard, ControlledField, LoadingModalState } from "@organisms";
 import { URL_REGEX, openInNewTab } from "@utils";
+import { useModal } from "@context";
 
 type StorageInformationProps = {
   setStep: Dispatch<SetStateAction<number>>;
@@ -29,6 +30,7 @@ export const StorageInformation = ({ setStep }: StorageInformationProps) => {
     onClickDownloadJson,
     isLoading,
   } = useCreateGovernanceActionForm(setStep);
+  const { openModal, closeModal } = useModal<LoadingModalState>();
   const { screenWidth } = useScreenDimension();
 
   const fileName = getValues("governance_action_type");
@@ -44,6 +46,21 @@ export const StorageInformation = ({ setStep }: StorageInformationProps) => {
   useEffect(() => {
     generateMetadata();
   }, []);
+
+  useEffect(() => {
+    if (isLoading) {
+      openModal({
+        type: "loadingModal",
+        state: {
+          title: t("createGovernanceAction.modals.loading.title"),
+          message: t("createGovernanceAction.modals.loading.message"),
+          dataTestId: "storing-information-loading",
+        },
+      });
+    } else {
+      closeModal();
+    }
+  }, [isLoading]);
 
   return (
     <BgCard